+++ R/data.R
@@ -22,7 +22,11 @@
 #' \item{matvals}{The name of a column in a tidy data frame containing matrices.}
 #' \item{matnames}{The name of a column in a tidy data frame containing matrix names.}
 #' \item{sector}{The name of a column in a tidy data frame containing sector names.}
+#' \item{country}{The name of a column in a tidy data frame containing names of countries.}
+#' \item{year}{The name of a column in a tidy data frame containing years.}
+#' \item{method}{The name of a column in a tidy data frame containing the method for accounting for the primary equivalent of renewable electricity.}
 #' \item{energy_type}{The name of a column in a tidy data frame containing names of energy types.}
+#' \item{last_stage}{The name of a column in a tidy data frame last stages of the energy conversion chain.}
 #' }
 #'
 #' @examples
@@ -67,6 +71,17 @@
 #' \item{eta_i}{The name of a column in a wide-by-matrices data frame containing industry efficiencies.}
 #' \item{eta_pfd_gross}{The name of a column in a wide-by-matrices data frame containing  efficiencies between the primary and gross final demand stages.}
 #' \item{eta_pfd_net}{The name of a column in a wide-by-matrices data frame containing efficiencies between the primary and net final demand stages.}
+#' \item{eta_pf}{The string name for primary-to-final efficiency.}
+#' \item{eta_fu}{The string name for final-to-useful efficiency.}
+#' \item{eta_pu}{The string name for primary-to-useful efficiency.}
+#' \item{eta_ps}{The string name for primary-to-services efficiency.}
+#' \item{eta_us}{The string name for useful-to-services efficiency.}
+#' \item{eta_pw}{The string name for primary-to-well-being efficiency.}
+#' \item{eta_sw}{The string name for services-to-well-being efficiency.}
+#' \item{efficiency_name_suffix}{The suffix for names of columns containing efficiency names.}
+#' \item{gross_net}{The name of a column that contains "Gross" or "Net" for the type of efficiency.}
+#' \item{gross}{The entry in the `gross_net` column that identifies a gross efficiency.}
+#' \item{net}{The entry in the `gross_net` column that identifies a net efficiency.}
 #' }
 #'
 #' @examples