+++ data-raw/create_constants.R
@@ -26,7 +26,11 @@ psut_cols <- list(resources = "R",
                   matnames = "matnames",
                   matvals = "matvals",
                   sector = "Sector",
-                  energy_type = "Energy.type")
+                  country = "Country",
+                  year = "Year",
+                  method = "Method",
+                  energy_type = "Energy.type",
+                  last_stage = "Last.stage")
 usethis::use_data(psut_cols, overwrite = TRUE)
 
 
@@ -52,7 +56,18 @@ usethis::use_data(aggregate_cols, overwrite = TRUE)
 
 efficiency_cols <- list(eta_i = "eta_i",
                         eta_pfd_gross = "eta_pfd_gross",
-                        eta_pfd_net = "eta_pfd_net")
+                        eta_pfd_net = "eta_pfd_net",
+                        eta_pf = "eta_pf",
+                        eta_fu = "eta_fu",
+                        eta_pu = "eta_pu",
+                        eta_ps = "eta_ps",
+                        eta_us = "eta_us",
+                        eta_pw = "eta_pw",
+                        eta_sw = "eta_sw",
+                        efficiency_name_suffix = "_name",
+                        gross_net = "GrossNet",
+                        gross = "Gross",
+                        net = "Net")
 usethis::use_data(efficiency_cols, overwrite = TRUE)